requirements.txt
flask
requests
openai
python-dotenv
mysql-connector-python
gunicorn               # si despliegas con Gunicorn
